Adds {example} blocks.
Output the contents of the block as well as the result of evaluation,
includes the `STDOUT`/`STDERR` streams.

Named `{example <name>}` blocks share the same sandbox module to allow
for commentary between parts of an example while retaining defined
values.

`{example}` and `{example <name>}` do not share state across pages
since the order of evaluation of pages is purposefully undefined.

(Currently an experimental feature and undocumented.)