Implement Selector interface
Currently the control flow in Documenter used to determine how
to transform a `Document` is defined in a rather cryptic way.

The "selectors" added here should hopefully allow for a more
straight forward approach to implementing the "pipeline" tuple
and "exander" functions. Neither of those are done in this commit,
but will be added in a subsequent one.