Cross-version compatible docsystem interface
Adds a `DocSystem` module that provides a programmatic interface to
the Julia docsystem rather than having to rely on `at-doc` and
concatenated `MD` objects.

Main interface is `DocSystem.getdocs` which takes a `Binding` and
optional type signature. Keyword arguments for the signature comparision
and which modules to search are also provided. `getdocs` returns
a `Vector{DocStr}` containing matching docstrings.

The differences in `0.4`'s internal representation are handled
automatically and shouldn't be visible to the caller. Some metadata
is only available in `0.5` since it wasn't captured in `0.4`.