Use multiple dispatch to determine the Writer
The writers need to be able to implement render methods on a whole
Document, instead of just on a Page-by-Page basis, since generally
all formats have to write some generic files as well, not just
individual pages.

Effectively this change means that every writer should overload a
    render(::Format, doc)
method, which should then perform the whole write process, including
all generic, non-page files. This overload should be the only method
that the rest of the Documenter code should be aware of.