Clean up page names
Currently the keys of Document's .internal.pages are paths to the
source files. They should instead be systematic page names. This way
they can be used for constructing paths etc.

In practice this means that we start with a path relative to the source
directory and strip the .md extension from the end.

E.g. `src/dir/file.md` becomes `dir/file`