Move populate! calls to a separate Builder step
Currently the writer populates the IndexNodes and ContentsNodes with
links, but that should be a separate step. This separates the populate!
calls into a separate Builder.Populate step, which runs just before
the Builder.RenderDocument.

This fixes the issue of duplicate links in the Index/ContentsNodes when
Builder.RenderDocument gets called multiple times after the document has
been constructed by the previous steps in the pipeline.