Use a `Pipe` for output redirection
This returns to the original method of output redirection of only
changing streams without the `withoutput` `do` block rather than
globally in the `Document` object with some changes.

No `at-async` to combine the output streams. To deal with
`STDOUT`/`STDERR` interleaving we use a `Pipe` and call
`Base.link_pipe` (unexported) to redirect both `STDOUT` and `STDERR`.