Fix index ordering when extension changes
Currently the file extensions are not changed in `pagesmap`s, but are
in the generated lists of `IndexNode`s and `ContentsNode`s. This
confuses the sort in the `populate!` functions when the extension is not
`.md` anymore. With this patch the extension change is postponed to the
writing step, making the `IndexNode` and `ContentsNode` sort
format-independent.

Fixes #226.