Add first-level headings to navmenu
Add first-level headings to the navigation menu. The only exception is
when the very first block of a page is a first-level heading -- in this
case that particular heading is ignored. This is consistent with how
the headings are usually used, with the first heading being the page
title.

The `<li>` tags of the first-level headers have the class `.toplevel`
attached to them, so that we can style them in the CSS.

Fixes #225.

(cherry picked from commit dca757dcb940500beaed9e70dab002f32adc6b24)