Move test/examples build code
Move the code of setting up the modules and building the examples docs
under examples/ to test/examples/make.jl. This allows the examples/ to
be built without calling the whole runtests.jl file. The pages/* are
considered to be part of the examples/.

The Mod and AutoDocs modules are now indented to help with readability.

Also, the output from the examples/ builds will now be stored under
test/examples/builds/<format>, instead of test/examples/build-<format>.