Deprecate the `Format` enum
Use a `Symbol` instead so that we can allow for additional 'plugin'
formats to be added from outside the package; for example if the
format has large dependencies that we don't want to include in
Documenter itself.

`.user.format` is now stored as a `Vector{Symbol}` so that we can,
in future, build several formats at once without needing to rebuild
the entire doctree for each one. Currently only a single format option
is supported, multi-format will have to wait until the doctree refactoring
PR is revisited.