Remove 'documentation' from page titles
Not all sites built using Documenter are going to be specific to
"documentation". Removing the word allows us to cater to users how
might just be writing a tutorial or organisation page and so might
not want to be labelled as documentation. Getting the old naming
back is as simple as adding ` documentation` to the `sitename`.