Update DifferentialEquations.jl documentation link
It was moved to a separate repository, DiffEqDocs.jl, to make it easier to manage the docs for the whole org. It's probably a good example for how to setup travis to do so too!
(cherry picked from commit 5c5b57a7e8889980880ae3e7f1f7b3860aac15f5)