Update the version selector
Add a version argument to makedocs that will allow specifying the
version shown in the selector by default. If it is not specified then
the version selector will be empty and have visibility: hidden so that
an empty selector wouldn't show up. Using visibility: instead of
display: so that the sidebar wouldn't jump when the selector gets
enabled.

Separate the version selector javascript into a separate requirejs block.
Amend the version selector javascript so that existing selector options
get checked and updated, if necessary. A siteinfo.js file will now also
be generated in the root of the build when being deployed which can be
used to determine the current version.

Move version selector out of search <form> tag, making it easier to
style and since it is not conceptually related to the search.

Update the style of the selector to be uniform across browsers, more in
line with the style and more connected to the package name rather than
the search box. The SVG arrow is necessary to make the dropdown
indicator uniform.