Fix doctest highlighting
There was a confusion regarding how the Markdown.Code language attribute
was handled for doctests depending on whether they were being run or
not. If the doctests do run, then the code blocks get a `julia` or
`julia-repl` (previously `jlcon`) language attribute. However, if the
doctests are skipped then they blocks retain the `jldoctest` attribute.

The *Writer modules therefore have to check for this -- previously
either `jlcon` or `julia` was applied in the HTML build, depending on
whether the doctests were enabled or not.