Switch all .js dependencies to cdnjs.com
MathJax CDN in shutting down, so we need to change that to cdnjs.com.
Additionally, it would probably be best to have all the dependencies
come from the same CDN, so we'll also change the CDN for jQuery, and
give it a patch-version boost while we're at it.

Fixes #473, #479.

(cherry picked from commit c240596d8a356f6f88c09ad0e567662aa6ad0399)