Determine extensions in Writers
In preparation of pretty URLs, make file extension mangling a
responsibility of the particular Writer module. Before the writing step,
all links should point to the original files, which generally have the
.md extension.

This only affects LatexWriter, since HTMLWriter already fixes .md links
and for the :markdown format nothing is actually done to the links in
the cross-referencing phase.