update to latest lunr.js version
For lunr.js changes see
https://github.com/olivernn/lunr.js/blob/master/CHANGELOG.mdown
For upgrade guide see https://lunrjs.com/guides/upgrading.html

The index is now immutable, hence the documents have to be added before
the end of the configuration function.

Build time field boosting is no longer supported, and replaced by query
time boosting. In this commit it is excluded. Before adding it again,
evaluate if it is still required, because with default settings it
should score the short titles higher anyways.