Replace Nullable{T} with Union{T, Void} (#576)
Nullable is going to be removed in Julia 0.7. The replacement is Union{T, Void}
or Union{Some{T}, Void}. The latter allows distinguishing a valid 'nothing' from
the absence of a value, but it is not yet available and does not really appear
to be required, except maybe for DocStr.object (see TODO).