HTMLWriter: add html_canonical option (#622)
* HTMLWriter: add html_canonical option

This allows search engines to point their users only to the most
relevant version of a certain page. This is particularly useful
when serving many similar pages (e.g. many versions of the same
documentation).

See https://en.wikipedia.org/wiki/Canonical_link_element for details.

* Remove superfluous parameter combinations in unit tests

It makes sense to have every parameter take on every value (e.g.
with and without `html_canonical`) but there's no need for a
combinatorial explosion of combinations of unrelated parameters.

Thanks to @mortenpi for suggesting this.

* Enable the new html_canonical feature for our own docs