Fix Markdown format test
Use correct capitalization for the assets/mkdocs/Documenter.css file.
The needed a bit of reorganization -- it was working before because it
was building into a dirty docs/build directory. Now the Markdown build
goes into a completely different directory. Also, no need to remove the
output files, instead let's clean the directory when the build starts.