Fix spurious Package.Package missing warnings
It seems that going from Julia 0.6 to 0.7 the way module bindings are
stored in the docs metadata changed from Binding(Mod, :SubMod) to
Binding(Mod.SubMod, :SubMod). Documenter, however, still constructs the
former Binding objects for modules that are listed just as SubMod in an
at-docs block.

This fixes the spurious errors that causes when checking for missing
docs by explicitly normalizing the module bindings that Documenter
creates before it crosschecks them against the bindings in docs
metadata.