Fix strict=true not failing for doctest failures (#1014)
The issue was introduced in #958, where the push!, which was hidden away
in the DocTests.report function, got accidentally removed. It is not
clear from its name that the report function should have a side-effect
like this, so it is cleaner if the push! calls are next to the report
calls, even if it creates slight code duplication.

Fix #1003