Build non-pages pages too in HTMLWriter
They can then be reached via their URL, which is determined from the
filename, as for the other pages. But they will not show up in the
sidebar nor in the bottom navigation.

If there is no filter applied, these pages may still show up in an
at-contents block. In fact, this actually fixes a bug where, in the HTML
output, an at-contents block would list such pages, even though the
writer does not produce any `.html` files, leading to broken links in
the output.