Handle invalid syntax in doctests (#1062)
* Handle invalid syntax in doctests

* Pass empty backtrace for parse errors

In the terminal, parse errors do not have a backtrace. But we still need
to initialize the result.bt field.

* Handle backtrace more properly

The heuristic in error_to_string which removes the backtrace below the
highest Core.eval call works in most cases but will fail if
makedocs/doctest gets called in a more complicated context, as there may
sometimes be eval calls higher up in the stacktrace.