Also handle \xor in PDF monospace
DejaVu Sans Mono does not render the \xor symbol, so we actually create
our own fallback system for this case where instances of this character
get replaced by the \unicodeveebar command in the TeX source.
\unicodeveebar then renders this character with DejaVu Sans.

In principle, this is pretty easily extensible to other missing Unicode
characters as well, should they come up.