Check env vars for non-Travis CI platforms
Special cases are now Travis, Gitlab, Drone, Cirrus and Appveyor. 

Not all of these CI services provide the same information as Travis 
does, or in the same way, so some massaging of variable contents is 
required.  Appveyor is arguably one of the worse ones here.

This is untested on most of these CI systems at the time this was 
committed; hence, any feedback is appreciated.