Add docs
So, the DocStringExtensions' $EXPORTS in the Documenter module docstring
creates at-ref links, so we need to make sure that we have docstrings
available to all exported names.

But you can't add docstrings in the at-enum, so I'm doing it separately.

Also, add all the refs to those variables while we're at it.