Add fallback Markdown -> Markdown2 convert methods
It's easy to end up with Symbols, Ints or other arbitrary Julia objects
in Markdown.MD trees. While it usually indicates a problematic docstring
or a parser bug, this makes sure that we do not just crash in those
situations. Rather, we just convert them to text nodes with repr().