Add fallback Markdown -> Markdown2 convert methods (#1075)
* Add fallback Markdown -> Markdown2 convert methods

It's easy to end up with Symbols, Ints or other arbitrary Julia objects
in Markdown.MD trees. While it usually indicates a problematic docstring
or a parser bug, this makes sure that we do not just crash in those
situations. Rather, we just convert them to text nodes with repr().

* Fix missing method error when re-using docstrings

Using the `@doc @doc(foo) bar` pattern wraps the docstrings in another
layer of Markdown.MD, which needs to be unwrapped before converting to
Markdown2.

* Add fallback rendering to mdconvert too

As the Markdown parser technically allows you to have arbitrary Julia
expressions in the Markdown tree, this makes sure that we do not crash
with a missing method error in that situation.