Fix problems with at-index/contents in LaTeXWriter
The \begin{itemize} / \end{itemize} were not correctly paired in the
LaTeX generated for at-contents blocks when the items went for than two
levels deep. Also, empty at-index and at-contents emitted an empty
\begin/\end pair, which errors, as you need at least one \item.

When storing the LaTeX build files with DOCUMENTER_LATEX_DEBUG, you can
now set it to a directory name, so that it would not generate a new
directory every time.

Also, except for one file, test/examples is not built with LaTeXWriter
on CI too in the new build stage, fixing #950.