Use 'Lato Medium' as the primary font
Lato from cdnjs provides separate font files for each weight, whereas
with Google Fonts we only source the one for font-weight: 400. The cdnjs
equvalent to that is using the 'Lato Medium' font which only load the
fotn file for font-weight: 400, more or less re-creating the look we had
with the font from Google Fonts.