Support prerendering of code blocks. (#1627)
Passing `prerender=true` to `Documenter.HTML` now renders syntax
highlighting of code blocks when building the page rather than
in the browser when the page is loaded. This requires a `node`
(NodeJS) executable in PATH or that a path to an executable is
passed as the `node` keyword argument. The `highlightjs` keyword
argument can be used to give a (local) file path to a custrom
hightlight.js library.

Example configuration:

  using Documenter, NodeJS_16_jll

  makedocs(;
      format = Documenter.HTML(
          prerender = true,            # enable prerendering
          node = NodeJS_16_jll.node(), # specify node executable (required if not available in PATH)
          # ...
      )
      # ...
  )

This feature is marked as experimental for now.