Use DocumenterTools as a dev dep in docs/
This does a few things:

1. The source links to the DocumenterTools docstrings do not work right
now because we Pkg.add it, but we need the source files to be located in
a Git repository for gitremote to determine the remote repository. We
can work around that by making it a Pkg.develop dependency instead.

2. If we add it as a dev dependency, we can use unreleased branches
which declare support for newer Documenter versions, and so we can use
the -DEV versions in the Documenter repository now. However, this
requires checking out a particular branch (or tag) of DocumenterTools.
Pkg.develop can not be used for this since it does not know how to check
out specific branches, and will fail as it tries to check out master
(which will not be compatible with e.g. 0.28.0-DEV).

3. Since adding it as a local dev dependency is a slightly non-trivial,
this creates a separate docs/instantiate.jl script which does the
necessary cloning and Pkg environment manipulations, just to make life
easier locally and in the CI config.