Remove unused code paths related to DocsNode
The DocsBlocks Expander step makes sure that the .docstr field of a
DocsNode is an MD object and that it has the :results metadata set.
Hence the if's are unnecessary and, in the LaTeXWriter case, the other
branch has been calling an non-existent method.

This patch removes those code paths and replaces the `if` with asserts,
just in case.