latexmk: use -interaction=batchmode and -halt-on-error
-interaction=batchmode is more appropriate than
-interaction=nonstopmode, since the latter includes "diagnostic
messages" that are generally just noise, and actually make it harder to
debug build failures.

-halt-on-error should help in the cases where LaTeX doesn't clearly
state that something was an error (like when you have missing fonts), by
aborting the build as soon as it encounters an error, and so the error
should be at the end of the logs regardless.

(cherry picked from commit 3968745d55cadc5244156edcb0164b5054a41a63)