ci: run in runci/* branches
This makes it easier to run CI on user branches, which is useful when
pushing to master without the noise of a pull request, but where it
would still be a good idea to run CI on the changes before adding the
commits to the master branch. Having these special runci/* branches
available in the CI settings means that it is not necessary to e.g.
cherry pick some commits or make temporary edits to enable CI on a user
branch (we still want to avoid CI on most branches, since most changes
will be merged via PRs).