Support regex => substitution pairs in doctest filters (#1989)
This patch adds support for using regex => substitution pairs in doctest
filters. This is useful when you do not want to replace the full match,
but only remove some part. As an example, consider a test where the
output may differ in the third decimal:

````
```jlddoctest
julia> 1.124
1.123
```
````

With this patch it is still possible to test the first 2 decimals using
the filter `r"(\d\.\d\d)\d*" => s"\1"`.