Retrigger precompilation when Project.toml changes
This patch make precompilation dependent on the Project.toml file. This
is necessary since the Documenter version is read from this file during
precompilation, and it will not be updated correctly otherwise. This
only matters if you are currently developing the package and will have
no effect on regular use/users.