Change the changelog infrastructure (#2184)
This patch changes how the changelog is written in order to support
links in both the CHANGELOG.md as viewed on GitHub, and in the generated
HTML. The changelog related code is moved into docs/changelog.jl which
now has two purposes: i) generate the Documenter friendly markdown for
the manual, and ii) generate links to add to the footer of the
CHANGELOG.md file directly. The latter can be triggered by 'make
changelog'.