Revise Plugins API, add `plugins` keyword arg for `makedocs` and `doctest` (#2249)
* Add test for plugins API

* Make list of `plugins` a keyword arg

Instead of passing plugin objects to `makedocs` as positional arguments
(which was undocumented), they are now passed as a a keyword argument
`plugins`.

The docstrings of the `Plugin` type and the `getplugin` function have
been corrected. They were at best misleading, and at worst flat-out
wrong.

* Accept `plugins` argument for `doctest`

Forward keyword arguments from `doctest` to `makedocs`. This
includes the `plugins` keyword argument, but potentially other keyword
arguments as well.