docs: explicitly note previews aren't cleaned up and fix up markdown (#2352)
* docs(hosting): fix up markdown

* Make the emboldened "Cleaning up `gh-pages`" section a heading so it's
more easily visible and the section is indexable.

* Fix the markdown syntax so that the link pointing to the GitHub docs
renders properly.

* docs(deploydocs): fix typo for `subfolder` kwarg (`s/folder/subfolder`)

The `subfolder` kwarg is displayed as `folder` in the docstring. Fix
this typo.

* docs(deploydocs): add note stating previews aren't auto cleaned up

Add a note to the `deploydocs` docstring explicitly stating that the
previews generated aren't automatically cleaned up and link to the
respective section of the documentation that has a GHA workflow for the
same.