Support scheme definition for GitLab remotes
GitLab is fairly easy to self-host. Although best practice, this also
means that the scheme over which GitLab is served is not guaranteed to
be `https`. For this reason, the provided `host` should be allowed to
define what scheme the GitLab instance is being hosted under and respect
that in generated links.

The same applies to ports and potentially paths, but those are already
supported given the current mechanism (provided no trailing slash is
included in the `host`).

One benefit of this approach is that it also allows a full "base" URL to
be provided as the host, e.g. the `CI_SERVER_URL` predefined variable
within GitLab CI environments[^1].

According to RFC3986[^2], URI schemes should be matched case
insensitively, which this approach does. It also prescribes that only
lowercase schemes should be produced, which this chooses to not do to
keep the implementation simpler.

[^1]: https://docs.gitlab.com/ee/ci/variables/predefined_variables.html
[^2]: https://datatracker.ietf.org/doc/html/rfc3986#section-3.1