Caution against dynamic `prettyurls`
We should steer people away from using

    prettyurls = get(ENV, "CI", nothing) == "true"

(or even `prettyurls = false`) and towards using `LiveServer`.

It is best to have a consistent setting and to always use a webserver
when viewing documentation locally. This avoids situations where there
might be subtle difference with `prettyurls=true` or `prettyurls=false`
and the documentation might work correctly when viewed locally, but be
subtly broken (without an error message!) when deploying on GitHub
Action. One such situation is a `@raw` block that references a local
image. The correct relative path of that image would be different
depending on the `prettyurls` setting.