Resolve `@ref` bindings in current module _and_ Main
For any `@ref` link in a docstring to a code object, Documenter
currently tries to find a binding for the code in the module where the
docstring is defined. This often requires importing symbols only because
they appear in a `@ref`. In some cases, this is not even possible, as it
would create circular references.

With this change, Documenter now tries to find binding for `@ref` links
in docstrings first in the local module, and then in `Main`, which is
basically `docs/make.jl`. Thus, anything imported in `docs/make.jl` can
be referenced anywhere.