Resolve `@ref` bindings in current module and `Main` (#2470)
For any `@ref` link in a docstring to a code object, Documenter
currently tries to find a binding for the code in the module where the
docstring is defined. This often requires importing symbols only because
they appear in a `@ref`. In some cases, this is not even possible, as it
would create circular references.

With this change, Documenter now tries to find binding for `@ref` links in
docstrings first in `CurrentModule`, and then in `Main` (that is,
`docs/make.jl`) *if and only if* the link is a fully qualified name. Thus,
anything imported in `docs/make.jl` can be referenced anywhere, without allowing for crazy non-obvious references.