Tone down admonition coloring and adjust border-width of elements
This patch tones down the admonitions by removing the background color
in the admonition header and the lighter background color in the body.
Instead the text in the admonition header is colored, and the colored
border thickness increased slightly.

The border width of code blocks and docstrings have been increased to
match the new border width of admonitions. Code blocks and docstrings
now also have a (non-zero) border radius to match admonitions (which
already were rounded before this patch).