When fixing doctests with errors, don't overwrite `[...]` (#2642)
* When fixing doctests with errors, don't overwrite `[...]`

Sometimes it is useful to show errors in doctests. However if those
contain stacktraces, this can be problematic as those contain line
numbers which will change over time (and often also when e.g. Julia or
some dependencies change). And sometimes it is simply desirable to
shorten the error message for other reasons.

For this reason Documenter supports replacing part of the error message
with `[...]` which can be used to omit e.g. a stacktrace while still
passing doctest checks.

However so far using Documenter's `doctest=:fix` feature would replace
all uses of `[...]` by the full error message, making it unnecessarily
tedious to update doctests in packages that use this feature.

This patch changes that by skipping doctest fixups for tests where the
reference output ends with `[...]` and otherwise matches the actual
error message as a prefix.

* add issue references

---------

Co-authored-by: Morten Piibeleht <morten.piibeleht@gmail.com>