Show file paths relative to cwd in messages (#2659)
This is useful for anyone using a terminal emulator that makes
file paths with line numbers clickable (to open an editor at that
file and line), such as VS Code or iTerm and many others.

It also matches the behavior of most UNIX tools, e.g. `git status`
or compiler error messages etc. all tend to be relative to the
current working directory.