Guide users to use workspaces when setting up Documenter
Now that Julia v1.12 is out, we can all have the nice things that the
Pkg folks have been cooking up for us.  This is an initial draft at
guiding users to use `[workspace]` in their `Project.toml`.  I can't
help but feel that having a one-shot command to setup Documenter for
this would be nicer, but c'est la vie.